Speed upgrades

+++ src/main/java/jaminv/advancedmachines/lib/machine/MachineTile.java
@@ -11,11 +11,11 @@ public interface MachineTile extends CanProcess, RedstoneControlled {
 
 	/** Allows the machine to process more quickly. 
 	 * Called every time a new process is started. */
-	public default int getSpeedMultiplier() { return 1; }
+	public default float getSpeedMultiplier() { return 1; }
 	
 	/** Allows the machine to output more secondary items.
 	 * Called every time a new process is started. */
-	public default int getProductivityMultiplier() { return 1; }
+	public default float getProductivityMultiplier() { return 1; }
 	
 	public void onControllerUpdate();
 }
